Deprecate .mjs extension
